Open branch new file structure
