+++ include/drivers/modem/modem.h
@@ -1,0 +1,13 @@
+#ifndef MODEM_H_
+#define MODEM_H_
+
+#include <stdint.h>
+#include <string.h>
+#include <stdio.h>
+#include "errornum.h"
+#include "stm32_config.h"
+#include "at_commands.h"
+
+uint8_t rc7620_init(void);
+
+#endif
